+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Added method to add a new account to the wallet object.
 - Added method to export the private key of an address
 - Added importWallet() to import account using privateKey.
+- Added getBalance() method